backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

SQLAlchemy>=2.0.30
psycopg[binary]>=3.2.1
alembic>=1.13.2

pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9

python-jose[cryptography]>=3.3.0
passlib>=1.7.4

email-validator>=2.1.1

